Adde more controllers and updated routes
