+++ app/Http/Controllers/CustomerController.php
@@ -21,6 +21,32 @@ class CustomerController extends Controller
 
              return Response::make('User created! Hurray!');
         }
+		
+		public function showcustomeroverview(){
+          return view('customercontent.customer-overview');
+  }
+
+  public function showcustomermenuoverview(){
+          return view('customercontent.customer-menuoverview');
+  }
+
+  public function showcustomerconfirmation(){
+          return view('customercontent.confirmationpage');
+  }
+
+  public function showcpeditaddress(){
+          return view('customercontent.customer-profile-editaddress');
+  }
+
+  public function showcpeditorders(){
+          return view('customercontent.customer-profile-editorders');
+  }
+
+  public function showcustomerprofile(){
+          return view('customercontent.customer-profile');
+  }
+		
+
