+++ app/Http/Controllers/RestaurantController.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+use App\Http\Requests;
+
+class RestaurantController extends Controller
+{
+     public function showrestauranthistory(){
+          return view('restaurantcontent.restaurant-history');
+  }
+
+  public function showrestaurantmadmin(){
+          return view('restaurantcontent.restaurant-menuadmin');
+  }
+
+  public function showrestaurantmedit(){
+          return view('restaurantcontent.restaurant-menuedit');
+  }
+
+  public function showrestaurantmoverview(){
+          return view('restaurantcontent.restaurant-menuoverview');
+  }
+
+  public function showrestaurantoverview(){
+          return view('restaurantcontent.restaurant-overview');
+  }
+
+  public function showrestaurantprofile(){
+          return view('restaurantcontent.restaurant-profile');
+  }
+
+  public function showrestaurantrestrictions(){
+          return view('restaurantcontent.restaurant-profile-restrictions');
+  }
+
+  public function showrestaurantprofilehours(){
+          return view('restaurantcontent.restaurant-profile-hours');
+  }
+}
